Separates favorites by user

+++ src/app/pages/article/article.component.ts
@@ -1,6 +1,6 @@
 import { Component, effect, inject } from '@angular/core';
 import { ArticleService } from '../../services/article.service';
-import { ActivatedRoute, RouterLink } from '@angular/router';
+import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { GalleryImageComponent } from '../../components/gallery-image/gallery-image.component';
 import { IlustratedLinkComponent } from '../../components/ilustrated-link/ilustrated-link.component';
 import { InfoboxSectionComponent } from '../../components/infobox-section/infobox-section.component';
@@ -8,6 +8,7 @@ import { firstValueFrom, map, switchMap } from 'rxjs';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { DatabaseService } from '../../services/database.service';
 import { IonToast } from '@ionic/angular/standalone';
+import { UserService } from '../../services/user.service';
 
 @Component({
   selector: 'app-article',
@@ -24,6 +25,8 @@ import { IonToast } from '@ionic/angular/standalone';
 export class ArticleComponent {
   articleService = inject(ArticleService);
   databaseService = inject(DatabaseService);
+  userService = inject(UserService);
+  rotuer = inject(Router);
   route = inject(ActivatedRoute);
   articleId$ = this.route.params.pipe(map(params => params['id']));
   article = toSignal(this.articleId$.pipe(switchMap(id => this.articleService.getArticle(id))));
@@ -32,10 +35,13 @@ export class ArticleComponent {
   isToastOpen = false;
 
   async toggleFavorite() {
+    if (!this.userService.getCurrentUser()) {
+      this.rotuer.navigate(['/log-in']);
+      return;
+    }
     let id = await firstValueFrom(this.articleId$);
-    let article = this.article()!;
     if (!this.isFavorite()) {
-      this.articleService.addToFavorites({id,title:article.title,image_url:article.articleImage});
+      this.articleService.addToFavorites(id);
       this.toastMessage = 'Favorite successfully added';
     } else {
       await this.articleService.removeFromFavorites(id);
